// src/model/courseclass.model.ts
export class CourseClass {
    id: string;
    title: string;
    description: string;
    status: string;
    start_date: string;
    end_date: string;
    course_id: string;
    days_course: number;
    description_course: string;
    description_courseclass: string;
    end_date_courseclass: string;
    image_course: Object;
    manual_course: string;
    start_date_courseclass: string;
    status_courseclass: string;
    title_course: string;
    title_courseclass: string;
}